0.11-stable: Check permissions when rendering wiki links so that resources not accessible to a user don't appear as a quickjump result.
More precisely, this changeset fixes attachment, ticket, milestone, changset and changelog links. Wiki links already had the permission check. Source browser links are '''not''' fixed, as the permissions to be checked depend on the resource type (file or directory).

The rendering of wiki links to resources that are either not existing or inaccessible is also changed to remove the `href=` attribute, as they would lead to an error page anyway. The exceptions are wiki page links (to allow creating the page) and milestones (which should grow the same functionality shortly).

Closes #7655.


git-svn-id: http://svn.edgewall.org/repos/trac/branches/0.11-stable@7604 af82e41b-90c4-0310-8c96-b1721e28e2e2