fixes genshi dependency
fixes genshi dependency for broken setuptools install.  setuptools
doesn't glean the revision of an svn checkout properly, causing it to
not be appended to the genshi distribution version.  Trac requires a
certain revision of genshi, and seince the revision number isn't
present, trac refuses to start.  The setuptools-fix.patch removes that
dependency.  This could cause version mismatches and odd behavior if
genshi isn't upgraded with trac.